// src/state/types.ts
export type SerializedBigNumber = string
export interface Investment {
  amount: string
  lockUntil: number
}
export interface Farm extends FarmConfig {
  tokenAmountMc?: SerializedBigNumber
  quoteTokenAmountMc?: SerializedBigNumber
  tokenAmountTotal?: SerializedBigNumber
  quoteTokenAmountTotal?: SerializedBigNumber
  lpTotalInQuoteToken?: SerializedBigNumber
  lpTotalSupply?: SerializedBigNumber
  tokenPriceVsQuote?: SerializedBigNumber
  poolWeight?: SerializedBigNumber
  lockingPeriod?: number
  fixedApr?: number
  totalAmount?: number
  userData?: {
    allowance: string
    tokenBalance: string
    stakedBalance: string
    earnings: string
    freeInvestmentAmount: string
  }
}